add --add-opens=java.base/java.nio=ALL-UNNAMED back to script
This is needed for snowflake driver at runtime. Specifying it this way (without the specific reference to the arrow module) does not generate a warning at startup.

This should support allow the snowflake driver to still funciton
https://community.snowflake.com/s/article/JDBC-Driver-Compatibility-Issue-With-JDK-16-and-Later

Though my snowflake trial account is expired so I'm not able to :100: confirm at the moment.